Decouple ONNX-GenAI metadata contracts
Move shared workflow contract primitives and deterministic YAML serialization out of the domain workflow builder. Rewire inference metadata and sibling producers to direct owners so inference metadata no longer imports workflow metadata, while preserving package exports and emitted bytes.
