+++ src/mobius/integrations/onnx_genai/__init__.py
@@ -34,6 +34,9 @@ The core model/task/component layers remain runtime-agnostic; all onnx-genai
 specific code lives here.
 """
 
+from mobius.integrations.onnx_genai._workflow_contract import (
+    add_policy_components_to_workflow,
+)
 from mobius.integrations.onnx_genai.auto_export import write_onnx_genai_config
 from mobius.integrations.onnx_genai.comfyui import (
     ComfyUIWorkflow,
@@ -55,7 +58,6 @@ from mobius.integrations.onnx_genai.decoder_metadata import (
 )
 from mobius.integrations.onnx_genai.inference_metadata import (
     SchedulerConfig,
-    add_policy_components_to_workflow,
     build_diffusion_pipeline_metadata,
     build_multimodal_pipeline_metadata,
     build_speech_to_text_pipeline_metadata,
