+++ src/mobius/integrations/onnx_genai/workflow_metadata.py
@@ -15,7 +15,6 @@ from collections.abc import Mapping, Sequence
 from typing import Any, ClassVar
 
 import onnx_ir as ir
-import yaml
 
 from mobius._constants import (
     STATIC_CACHE_KV_SEQUENCE_LENGTH,
@@ -109,17 +108,26 @@ from mobius.generation import (
     build_zeros_like,
     rotary_axis_count,
 )
-from mobius.integrations.onnx_genai.inference_metadata import (
-    _name_image_preprocessing_program,
+from mobius.integrations.onnx_genai._metadata_io import _dump_yaml
+from mobius.integrations.onnx_genai._workflow_contract import (
+    _cache_output_candidates,
+    _component,
+    _contract,
+    _effect,
+    _invoke,
+    _model_cache_pairs,
     _port,
+    _publish_workflow_v1,
+    _request_aligned,
     _shape_metadata,
+    add_policy_components_to_workflow,
+)
+from mobius.integrations.onnx_genai.inference_metadata import (
+    _copy_runtime_assets,
+    _name_image_preprocessing_program,
     _source_asset_path,
     add_adapter_service_to_metadata,
-    add_policy_components_to_workflow,
     build_native_vlm_package_metadata,
-    declare_input_admission,
-    declare_request_alignment,
-    request_batch_layout,
 )
 from mobius.integrations.onnx_genai.package_facts import (
     MEDIA_TOKEN_ROLES,
@@ -128,110 +136,12 @@ from mobius.integrations.onnx_genai.package_facts import (
     source_declared_value,
 )
 
-
-class _NoAliasSafeDumper(yaml.SafeDumper):
-    def ignore_aliases(self, data: Any) -> bool:
-        return True
-
-
 _CTC_TOKEN_ROLES = tuple(role for role in TEXT_TOKEN_ROLES if role.name == "pad_token_id")
-
-
-def _dump_yaml(metadata: dict[str, Any], handle: Any) -> None:
-    yaml.dump(metadata, handle, Dumper=_NoAliasSafeDumper, sort_keys=False)
 
 
 def _source_model_value(source: str | None, name: str, fallback: Any) -> Any:
     """Resolve a value from packaged runtime metadata when available."""
     return source_declared_value(source, name, fallback)
-
-
-def _contract(value: ir.Value) -> dict[str, Any]:
-    port = _port(value)
-    dtype = {"fp16": "float16", "bf16": "bfloat16", "fp32": "float32"}.get(
-        port.dtype, port.dtype
-    )
-    shape = _shape_metadata(port)
-    contract: dict[str, Any] = {
-        "dtype": dtype,
-        "rank": port.rank,
-        "shape": shape,
-    }
-    layout = request_batch_layout(shape)
-    if layout is not None:
-        contract["batch_layout"] = layout
-    return contract
-
-
-def _request_aligned(contract: dict[str, Any], axis: int = 0) -> dict[str, Any]:
-    """Mark a contract as carrying exactly one entry per in-flight request.
-
-    This is a structural batching fact, not a row identity: it tells the runtime
-    which axis to permute when it compacts the batch, while scheduler slots and
-    sequence handles stay runtime-private.
-    """
-    return {**contract, "batch_layout": {"kind": "request_aligned", "axis": axis}}
-
-
-# Translation between the port vocabulary this producer *mints* when it builds
-# a graph and the runtime's architecture-neutral role vocabulary. Both sides are
-# fixed vocabularies and Mobius owns one of them: the task builders in
-# ``mobius.tasks`` choose these exact names, so reading them back here is a
-# lookup, not an inference about a graph of unknown provenance. A port outside
-# this vocabulary carries no role, because a workflow that guesses is worse than
-# one that stays silent.
-_PORT_ROLES: dict[str, str] = {
-    "input_ids": "token_ids",
-    "inputs_embeds": "inputs_embeds",
-    "attention_mask": "attention_mask",
-    "position_ids": "position_ids",
-    "logits": "logits",
-    "last_hidden_state": "hidden_states",
-    "encoder_hidden_states": "encoder_hidden_states",
-    "audio_features": "audio_features",
-}
-
-
-def _component(
-    model: ir.Model,
-    artifact: str,
-    *,
-    effects: tuple[str, ...] = (),
-) -> dict[str, Any]:
-    """Declare one ONNX-backed workflow component: its artifact and port roles.
-
-    A component declares only what its artifact cannot say about itself. The
-    ``.onnx`` file is shipped inside the package and is authoritative for which
-    ports exist and what dtype, rank and shape each one has, so transcribing
-    that into YAML would create a second copy of a fact the package already
-    carries — one that can drift from the graph and that nothing cross-checks
-    at rest. The runtime resolves ports against the live session instead, which
-    catches a name the graph does not expose rather than agreeing with a stale
-    echo of it.
-
-    What no graph carries is what a port *means*. ``input_ids`` and
-    ``position_ids`` are both rank-2 ``int64``; nothing in the file says which
-    one is the autoregressive sequence. An invocation binds an SSA value to a
-    port, which records which value arrives but not whether it is tokens, a mask
-    or logits — and that second fact is what a runtime needs before it can
-    specialize a decode step. So ``roles`` is the whole declaration here.
-
-    Only ports in this producer's own vocabulary get a role, and state ports
-    never need one: the group that carries them already names its pairs, which
-    is also where the fixed-capacity scatter ABI is stated.
-
-    ``batch_capacity`` is intentionally absent. A request-aligned or dynamic
-    batch axis is structural shape information, not proof that co-batching
-    preserves each request's result. Builders may add that semantic permission
-    only after the complete grouped contract has been authored and validated.
-    """
-    del effects
-    named = [str(value.name) for value in (*model.graph.inputs, *model.graph.outputs)]
-    roles = {name: _PORT_ROLES[name] for name in named if name in _PORT_ROLES}
-    declaration: dict[str, Any] = {"implementation": {"kind": "onnx", "artifact": artifact}}
-    if roles:
-        declaration["ports"] = {"roles": roles}
-    return declaration
 
 
 def _declare_batch_capacities(
@@ -435,205 +345,6 @@ def _grammar_adapter_component(action: str) -> dict[str, Any]:
         # ABI when the batch changes; without it the FSM rows would drift out
         # of correspondence with the sequences they guide.
         "row_scope": {"axis": 0, "stateful": True},
-    }
-
-
-def _effect(consumes: str, produces: str) -> dict[str, str]:
-    return {"consumes": consumes, "produces": produces}
-
-
-def _publish_workflow_v1(workflow: dict[str, Any]) -> dict[str, Any]:
-    """Publish structured steps and logical carries without compiler bookkeeping."""
-    graph = workflow.pop("graph")
-    workflow.pop("initial_effects", None)
-    for declaration in workflow.get("inputs", {}).values():
-        source = declaration.get("source")
-        if isinstance(source, dict) and source.get("kind") == "request":
-            source.pop("field", None)
-
-    # Every workflow value whose leading dimension is the batch symbol holds one
-    # entry per in-flight request, so declare that structurally instead of leaving
-    # a runtime to infer it. Graph-derived contracts already carry the layout;
-    # this covers the hand-written declarations the runtime compares them against
-    # when it validates a carry, a binding, or an emit.
-    def _declare_row_alignment(contract: Any) -> Any:
-        if (
-            isinstance(contract, dict)
-            and "batch_layout" not in contract
-            and request_batch_layout(contract.get("shape")) is not None
-        ):
-            return _request_aligned(contract)
-        return contract
-
-    for section in ("inputs", "outputs", "state"):
-        for declaration in workflow.get(section, {}).values():
-            declaration["contract"] = _declare_row_alignment(declaration.get("contract"))
-    for component in workflow.get("components", {}).values():
-        for side in ("inputs", "outputs"):
-            ports = component.get("ports", {}).get(side)
-            if not ports:
-                continue
-            for port, contract in ports.items():
-                ports[port] = _declare_row_alignment(contract)
-    substitutions: dict[str, str] = {}
-    loop_index = 0
-    cell_aliases = {
-        cell: f"{cell}_state" if cell in workflow.get("outputs", {}) else cell
-        for cell in workflow.get("state", {})
-    }
-    if any(cell != alias for cell, alias in cell_aliases.items()):
-        workflow["state"] = {
-            cell_aliases[cell]: declaration for cell, declaration in workflow["state"].items()
-        }
-
-    def collect_carried(node: dict[str, Any]) -> None:
-        if node["kind"] == "loop":
-            for carry in node.get("carried", []):
-                alias = cell_aliases.get(carry["cell"], carry["cell"])
-                substitutions[carry["body_input"]] = alias
-                substitutions[carry["next"]] = alias
-            collect_carried(node["setup"])
-            collect_carried(node["body"])
-        elif node["kind"] == "sequence":
-            for child in node["nodes"]:
-                collect_carried(child)
-        elif node["kind"] == "branch":
-            for case in node["cases"].values():
-                collect_carried(case)
-            if "default" in node:
-                collect_carried(node["default"])
-
-    def rewrite(value: Any) -> Any:
-        if isinstance(value, str):
-            return substitutions.get(value, value)
-        if isinstance(value, list):
-            return [rewrite(item) for item in value]
-        if isinstance(value, dict):
-            return {key: rewrite(item) for key, item in value.items()}
-        return value
-
-    def convert(node: dict[str, Any]) -> dict[str, Any]:
-        nonlocal loop_index
-        kind = node["kind"]
-        if kind == "sequence":
-            return {
-                "kind": "sequence",
-                "steps": [convert(child) for child in node["nodes"]],
-            }
-        if kind == "invoke":
-            return {
-                "kind": "invoke",
-                "component": node["component"],
-                "inputs": rewrite(node.get("inputs", {})),
-                "outputs": rewrite(node.get("outputs", {})),
-            }
-        if kind == "emit":
-            result = {
-                "kind": "emit",
-                "value": rewrite(node["value"]),
-                "output": node["output"],
-                "mode": node["mode"],
-            }
-            if "axis" in node:
-                result["axis"] = node["axis"]
-            if "valid_length" in node:
-                result["valid_length"] = rewrite(node["valid_length"])
-            if "when" in node:
-                result["when"] = rewrite(node["when"])
-            return result
-        if kind == "branch":
-            result = {
-                "kind": "branch",
-                "predicate": rewrite(node["predicate"]),
-                "cases": {name: convert(case) for name, case in node["cases"].items()},
-                "outputs": rewrite(node.get("outputs", {})),
-            }
-            if "default" in node:
-                result["default"] = convert(node["default"])
-            return result
-        if kind == "loop":
-            current_loop = loop_index
-            loop_index += 1
-            setup = node["setup"]
-            body = node["body"]
-            setup_steps = (
-                [convert(child) for child in setup["nodes"]]
-                if setup["kind"] == "sequence"
-                else [convert(setup)]
-            )
-            body_steps = (
-                [convert(child) for child in body["nodes"]]
-                if body["kind"] == "sequence"
-                else [convert(body)]
-            )
-            carried = []
-            for carry in node.get("carried", []):
-                cell = cell_aliases.get(carry["cell"], carry["cell"])
-                published_carry = {
-                    "cell": cell,
-                    "next": rewrite(carry["body_output"]),
-                }
-                initial = rewrite(carry["current"])
-                if workflow["state"][cell]["initializer"] != initial:
-                    published_carry["initial"] = initial
-                carried.append(published_carry)
-            active_cell = node.get("active_cell")
-            if active_cell is None:
-                active_cell = f"loop_{current_loop}_active"
-                active_initializer = f"package.{active_cell}"
-                workflow["inputs"][active_initializer] = {
-                    "contract": {"dtype": "bool", "rank": 1, "shape": [1]},
-                    "role": {"kind": "opaque"},
-                    "source": {"kind": "literal"},
-                    "required": False,
-                    "default": True,
-                }
-                workflow["state"][active_cell] = {
-                    "contract": {"dtype": "bool", "rank": 1, "shape": [1]},
-                    "scope": "invocation",
-                    "initializer": active_initializer,
-                    "recurrence": {"kind": "invariant"},
-                }
-                carried.append(
-                    {
-                        "cell": active_cell,
-                        "next": rewrite(node["condition"]),
-                    }
-                )
-            result = {
-                "kind": "loop",
-                "setup": setup_steps,
-                "steps": body_steps,
-                "continue_when": active_cell,
-                "max_iterations": rewrite(node["max_iterations"]),
-                "carried": carried,
-            }
-            if "termination" in node:
-                result["termination"] = node["termination"]
-            if "iteration" in node:
-                result["iteration"] = node["iteration"]
-            return result
-        raise ValueError(f"unsupported workflow node kind {kind!r}")
-
-    collect_carried(graph)
-    published = convert(graph)
-    workflow["steps"] = published["steps"] if published["kind"] == "sequence" else [published]
-    declare_request_alignment(workflow)
-    declare_input_admission(workflow)
-    return workflow
-
-
-def _invoke(
-    component: str,
-    inputs: dict[str, str],
-    outputs: dict[str, str],
-    _effects: dict[str, dict[str, str]] | None = None,
-) -> dict[str, Any]:
-    return {
-        "kind": "invoke",
-        "component": component,
-        "inputs": inputs,
-        "outputs": outputs,
     }
 
 
@@ -2309,38 +2020,6 @@ def write_audio_codec_workflow_metadata(pkg: Any, output_dir: str) -> str:
     with open(path, "w", encoding="utf-8") as handle:
         _dump_yaml(metadata, handle)
     return path
-
-
-def _model_cache_pairs(model: ir.Model) -> list[tuple[ir.Value, ir.Value]]:
-    outputs = {value.name: value for value in model.graph.outputs}
-    pairs = []
-    for past in model.graph.inputs:
-        present = next(
-            (
-                outputs.get(name)
-                for name in _cache_output_candidates(past.name or "")
-                if name in outputs
-            ),
-            None,
-        )
-        if present is not None:
-            pairs.append((past, present))
-    return pairs
-
-
-def _cache_output_candidates(past_name: str) -> tuple[str, ...]:
-    """Names an exporter may give the output that continues a cache input.
-
-    An appending cache renames ``past`` to ``present``; a static, indexed cache
-    keeps the buffer's name and prefixes the written result instead, because the
-    output is the same buffer rather than a longer one.
-    """
-    return (
-        past_name.replace("past_key_values", "present"),
-        past_name.replace("past.", "present."),
-        past_name.replace("past_", "present_"),
-        f"updated_{past_name}",
-    )
 
 
 def _constant_extent(dimension: Any) -> int | None:
@@ -7359,6 +7038,31 @@ def write_vlm_workflow_metadata(
     with open(path, "w", encoding="utf-8") as handle:
         _dump_yaml(metadata, handle)
     return path
+
+
+def write_native_vlm_package_metadata(
+    pkg: Any,
+    directory: str,
+    *,
+    config: Any,
+    source: str | None = None,
+    revision: str | None = None,
+) -> dict[str, str]:
+    """Write the published VLM package contract and its tokenizer/processor assets.
+
+    What lands in ``inference_metadata.yaml`` is the typed SSA workflow built by
+    :func:`build_vlm_workflow_metadata`, not the structural descriptor
+    :func:`~mobius.integrations.onnx_genai.inference_metadata.build_native_vlm_package_metadata`
+    returns.
+    """
+    # Assets first: the workflow document declares their package-relative
+    # locations under ``package.tokenizer.artifacts``, so they have to be in the
+    # package before the document that names them is written.
+    artifacts = _copy_runtime_assets(directory, source, revision=revision)
+    artifacts["inference_metadata"] = write_vlm_workflow_metadata(
+        pkg, directory, config, source=source
+    )
+    return artifacts
 
 
 def build_speculative_workflow_metadata(
